webstore/requirements.txt
Django==2.0.4
django-appconf==1.0.2
django-parler==1.9.2
django-statici18n==1.7.0
Pillow==5.1.0
pytz==2018.4
